Ajout titre en ascii, page about, interactions avec les boutons du menu highscores, ajout bouton pour afficher les high scores

+++ projets-qt/CodeAlex-non-qt-deso/src/option.cpp
@@ -16,7 +16,7 @@ bool option_assert_validity(const int & argc) {
 
 /* Corps de la procédure `option_parse_to` qui insère les arguments de lancement dans une map d'options */
 void option_parse_to(execoption & options, int argc, char** arg) {
-    for (size_t i = 1; i < argc; i += 2) {
+    for (int i = 1; i < argc; i += 2) {
         string op_name = string(arg[i]);
         string op_val = string(arg[i + 1]);
         options[op_name] = op_val;
